Add files via upload
Alot is starting to make sense. Sometimes =VLOOKUP formula works and sometimes i get #N/A. It makes me energised to foolproof the formula and i do a little dance when i get a correct value. If you know you know.Ofcourse Excel is superwide and am tackling it step by step(literally) and i find that is my learning pace. From data validation,to data filters,grouping and ungrouping, to adding images i am so psyched to unravel the all the things that Excel can do and apply it successfully in all my projects.